bot_question_reports/requirements.txt
pymongo==4.6.1
aiogram==3.4.1
python-dotenv==1.0.0
pydantic==2.5.2 